Extract Sinkhorn optimal transport to shared module
Consolidate duplicated Sinkhorn algorithm implementations from transweave.py
and weight_causal.py into a new optimal_transport.py module. This reduces
code duplication (~200 lines removed) and provides a single source of truth
for optimal transport utilities.
