+++ fabricpc/continual/transweave.py
@@ -38,127 +38,12 @@ from fabricpc.continual.config import (
     ShellDemotionTransWeaveConfig,
 )
 
-# ----------------------------
-# Core Sinkhorn Transport
-# ----------------------------
-
-
-def sinkhorn_transport(
-    cost: np.ndarray,
-    source_weights: Optional[np.ndarray] = None,
-    target_weights: Optional[np.ndarray] = None,
-    eps: float = 0.25,
-    iters: int = 15,
-    identity_bonus: float = 0.0,
-) -> np.ndarray:
-    """
-    Compute optimal transport plan using Sinkhorn algorithm.
-
-    Args:
-        cost: Cost matrix (n, m)
-        source_weights: Source marginal weights (n,)
-        target_weights: Target marginal weights (m,)
-        eps: Regularization parameter
-        iters: Number of Sinkhorn iterations
-        identity_bonus: Bonus for diagonal (identity) transport
-
-    Returns:
-        Transport plan (n, m) - doubly stochastic matrix
-    """
-    cost = np.asarray(cost, dtype=np.float64)
-    n, m = cost.shape
-
-    # Default to uniform weights
-    if source_weights is None:
-        source_weights = np.ones(n) / n
-    if target_weights is None:
-        target_weights = np.ones(m) / m
-
-    source_weights = np.asarray(source_weights, dtype=np.float64)
-    target_weights = np.asarray(target_weights, dtype=np.float64)
-
-    # Normalize weights
-    source_weights = source_weights / (np.sum(source_weights) + 1e-10)
-    target_weights = target_weights / (np.sum(target_weights) + 1e-10)
-
-    # Add identity bonus to diagonal (if square)
-    if n == m and identity_bonus > 0:
-        identity_cost = np.eye(n) * (-identity_bonus)
-        cost = cost + identity_cost
-
-    # Initialize Gibbs kernel
-    K = np.exp(-cost / (eps + 1e-10))
-
-    # Sinkhorn iterations
-    u = np.ones(n)
-    v = np.ones(m)
-
-    for _ in range(iters):
-        u = source_weights / (K @ v + 1e-10)
-        v = target_weights / (K.T @ u + 1e-10)
-
-    # Transport plan
-    transport = np.outer(u, v) * K
-
-    return transport
-
-
-def cosine_cost_matrix(
-    source: np.ndarray,
-    target: np.ndarray,
-) -> np.ndarray:
-    """
-    Compute cosine distance cost matrix.
-
-    Args:
-        source: Source representations (n, d)
-        target: Target representations (m, d)
-
-    Returns:
-        Cost matrix (n, m) where cost[i,j] = 1 - cosine_sim(source[i], target[j])
-    """
-    source = np.asarray(source, dtype=np.float64)
-    target = np.asarray(target, dtype=np.float64)
-
-    # Normalize
-    source_norm = source / (np.linalg.norm(source, axis=1, keepdims=True) + 1e-8)
-    target_norm = target / (np.linalg.norm(target, axis=1, keepdims=True) + 1e-8)
-
-    # Cosine similarity
-    similarity = source_norm @ target_norm.T
-
-    # Convert to cost (1 - similarity)
-    cost = 1.0 - similarity
-
-    return cost
-
-
-def euclidean_cost_matrix(
-    source: np.ndarray,
-    target: np.ndarray,
-) -> np.ndarray:
-    """
-    Compute squared Euclidean distance cost matrix.
-
-    Args:
-        source: Source representations (n, d)
-        target: Target representations (m, d)
-
-    Returns:
-        Cost matrix (n, m)
-    """
-    source = np.asarray(source, dtype=np.float64)
-    target = np.asarray(target, dtype=np.float64)
-
-    # ||s - t||^2 = ||s||^2 + ||t||^2 - 2*s.t
-    source_sq = np.sum(source**2, axis=1, keepdims=True)
-    target_sq = np.sum(target**2, axis=1, keepdims=True)
-
-    cost = source_sq + target_sq.T - 2 * (source @ target.T)
-    cost = np.maximum(cost, 0.0)  # Numerical stability
-
-    return cost
-
+# Import optimal transport utilities from shared module
+from fabricpc.continual.optimal_transport import (
+    sinkhorn_transport,
+    cosine_cost_matrix,
+    euclidean_cost_matrix,
+)
 
 # ----------------------------
 # Composer-level TransWeave
